Add query multitenancy and fix some bugs
This works for me end to end now. Since Grafana cannot set arbitrary
headers and we don't have an auth frontend in between Grafana and the
distributor yet, the distributor is temporarily taking the user ID from
basic auth. So to test with Grafana, edit the datasource and enable
basic auth with any user name (becomes userID) and password (is
ignored).

Switching around user IDs in basic auth in the data source seems to
indicate that it all works, end-to-end!