Remove obsolete IngesterError type
The use of this type was removed when the HTTP ingester client was
removed. Not sure if the 4xx passthrough as the result of the
distributor's HTTP response made sense in the first place, since I'd see
it as an internal server error from the user's point of view if we
encode something wrong to the ingester internally.