Implement per-user series limits
When a series limit is reached, samples for existing series should still
be stored, so this is treated as a partial failure.

I tried various approaches to error propagation and handling, including
more complicated ones, but went for this medium-simple approach in the
end that uses gRPC errors to communicate a series-exceeded error from
the ingester back to the distributor, still manages to only partially
fail for the affected series, and communicates the specific error back
to the user. Due to only returning the last error encountered when
writing to multiple ingesters, this may at times cover up the
series-exceeded error if there is another error afterwards.