Retry failed terminal flushes later in queue (#328)
Mitigates #316. Some chunks will fail to flush because of temporary reasons, others permanent. For permanent failures, we want to give other chunks a chance to flush, rather than blocking the whole queue on them. We do this by re-inserting the failed chunks at an arbitrary later point in the queue.