Fix mutation bug in ExtractMetricNameFromMatchers()
The passed-in matchers slice got mutated by the copies, leading to the
metric name being removed from the original matchers *iff* the metric
name was not the last element in the slice already. By pure luck, PromQL
only creates LabelMatcher lists where the metric name is the last
element. But in my remote query testing, I ended up constructing
matchers with the metric name somewhere in the middle. The result was
that the metric name was removed from the matchers and all kinds of
metrics (not just the ones for the right metric name) were being
returned for a query.