Compare chunks by key without memory allocations
The replacement is a bit long-winded, but very fast in operation.

The ordering will not be the same: previously we sorted lexically, so
0x10 would come before 0x9. But the Fingerprint is a hash so we aren't
bothered about the exact ordering, just that chunks in the same series
come out together in time order.

Also we would now compare the same chunk with and without checksums
differently, but given the key format changed at a point in time you
can't get the same key with and without checksum.