Increase histogram buckets to cover typical scenarios
The default max chunk age is 12 hours, so we really ought to cover
that in our buckets.

If the metric repeats exactly (all zeros, say, and with completely
regular timestamps) then we will can easily fit 12 hours in a 1K
chunk.  12 hours of 15-second samples is 2880 samples, so I extended
the length buckets to cover that.