Only serve DB-backed stuff if DB set
In #649, we only served DB-backed endpoints for ruler if the configs API flag
was _not_ set. This turns out to be hostile to a smooth migration, as we want
to have points where we are serving _both_ sets of endpoints so we can update
the service-ui to go from pointing at one to pointing at the other.