Use BigTable Columns as range keys
See https://github.com/weaveworks/cortex/issues/714 for motivation.

One thing to notice is that unlike existing schemas, here the structure
of the data being stored changed. This caused some backwards
incompatible changes as the `StorageClient` interface doesn't have any
time information in the `QueryPages` making it impossible to know if we
need to range over the row-key or the columns to retrieve the data.

For maintaining backwards compat, use `chunk.storage-client=gcpv1`.

Signed-off-by: Goutham Veeramachaneni <cs14btech11014@iith.ac.in>