Return chunks fetched by GetChunks on error (#791)
* Return chunks fetched by GetChunks on error

If DynamoDB is overloaded, GetChunks() may time out, but we should
return the chunks fetch so they can be put into the cache.  Then if
the client retries we will get a bit further next time.

* Make backoff parameters configurable

* Be able to defer mock error for a number of operations

So we can run a test where some operations succeed and then subsequent
ones fail.

* Refactor: extract function to create chunks for testing

* Add test for DynamoDB operation with partial result

Some kludging required to (a) make sure it runs only on DynamoDB
and (b) access the error parameters which are on a non-exported type.

Also lots of magic numbers.  It does the job.