Don't check the build image is up-to-date when not using it (#854)
* Don't check the build image is up-to-date when not using it

We had a bunch of rules in Makefile insisting the build image was
up-to-date in the `else` clause of `BUILD_IN_CONTAINER`, and a
matching bunch of CI instructions to `touch` the file. Lose all of
this.

* Put back one touch

The Makefile builds all images by default, so rather than creating a
more complicated make rule to exclude the build image, we fake it out
so it doesn't build.