Batch iterators.
The idea here is that instead of fetching 1 sample per iteration, fetch N.  This should reduce the amount of work we have to do in the heap etc.

The batches themselves are designs to be passed by value; the various helpers we use to merge them are designs to be used in such a way that doesn't require any allocations.

Signed-off-by: Tom Wilkie <tom.wilkie@gmail.com>