Remove tracing span for dynamoDB query
In most cases the spans for query() are one-for-one with the spans for
DynamoDB.QueryPages(), and when there are thousands of small queries
this just makes traces larger and harder to read.

To distinguish the case where a query has multiple pages from the case
where one page is retried, I added tags with the page number and retry
count.

Signed-off-by: Bryan Boreham <bryan@weave.works>