Make it nicer to test locally with minikube
This change modernizes the version of Prometheus used in the manifests
and updates the configuration to match. It also adds RBAC to run it on
modern k8s.

Finally, update the Makefile with a rule to make it easier to prime the
minikube Docker with images built locally and reflect the changes in
README.md.

Signed-off-by: Christian Funkhouser <cfunkhouser@heroku.com>