Make series IDs backwards-compatible again
PR #1377 caused an unforseen change in the formatting of label
strings, which in turn means that the SHA is different so the seriesID
changes.  It's more efficient to continue with the same IDs across an
upgrade.

I noticed there was an unused function `metricSeriesID()` and a test
for backwards-compatibility which would have caught this mistake, so
have pressed them back into service.

Signed-off-by: Bryan Boreham <bryan@weave.works>