Adds v11 schema to store label names within index.
Stores only label names and not the entire metric. Storing entire metric
will bloat the index by 30% and it doesn't really make sense to do it
right now. Adding just label names adds a tolerable 7% to the index.

Also, in Prometheus, we don't treat __name__ as a special label. We
always return it when calling /labels API and we should do the same
here.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>
Signed-off-by: Goutham Veeramachaneni <gouthamve@gmail.com>