querier: do not cache results if requested
Add an extra `Headers` field to the `PrometheusResponse` message which
contains the headers and their values that came from Prometheus.

Use them in other places to deduce if the response should be cached. If
`Cache-Control` is equal to `no-store` then it is *not* cached.

This will be used by the Thanos project to indicate when a partial
response has been returned. In such cases the result should not be
cached so that invalid data would not be stored there.

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>