runtimeconfig: initialize overridesReloadSuccess gauge in success state
There seem to be code paths where the overridesReloadSuccess gauge
gets initialized, but `NewRuntimeConfigManager()` never is called
(absence of both log lines: "failed to load config", "runtime config
disabled: file not specified").

promauto.NewGauge() seems to initialize the gauge with the value 0
(that seems to be common in Prometheus client libs).

In the current code, the value 0 is associated with the error
state, and the value 1 is associated with the success state.

Likewise, https://github.com/grafana/cortex-jsonnet/ defines
an alert that fires when this gauge is set to 0 for an extended
period of time.

That is, when `NewRuntimeConfigManager()` never gets called
the gauge indicates a false negative error state.

Starting in the success state is a pragmatic way to solve that
problem.

It might be better to use a counter for individual error or
-- if a gauge is preferred -- one could introduce a third
state referring to "initialized".

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>