runtimeconfig: register cortex_overrides_last_reload_successful dynamically
Only register the gauge when NewRuntimeConfigManager()
is called (when a runtime config manager is actually being
used).

Still, initialize it explicitly in state 1 (success)
instead of implicitly initializing it in state (error).

manager_test.go: subsequently the registry should
be used for testing the state of the metric in question.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>