pkg/ring/kv: add List and Delete operations
This commit extends the kv interface to add List and Delete operations.
These methods are not used within Cortex, but instead are meant to be
utilized by downstream consumers (e.g., grafana/agent).

Note that the memberlist KV implementation does not support deletes as
of this commit, but List has been implemented.

This commit also slightly tweaks the implementation of consul and etcd's
Get operation to return the same error when a key is not found. This is
useful for consumers to be able to detect a nonexistent key rather than
having to check for the specific error case of each individual
implementation.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>