Handle metric metadata in the ingesters
This change allow us to handle metric metadata sent by distributors within the Ingesters.

We use a in-memory map, were keys are the tenant ID and the values are another map with keys as metric names and values as metadata set. We only ever keep a single copy of each metadata alongside a timestamp of when was the last time we saw it. This is later used to purge the set of metadata we have not seen in a while and avoid the cache from ever-growing.

Signed-off-by: gotjosh <josue@grafana.com>