Allow starting ingester in JOINING state to auto join
If an ingester crashes during token transfer, or fails to rollback a
token transfer then it will start up with a ring entry in the JOINING
state. This will mean the ingester will no longer auto join because it
is not in the PENDING state, and cannot receive a transfer request
either. It will just stay in the JOINING state until manually deleted
and forgotten from the ring.

Instead, handle this case upon init, and set the state back to PENDING.
This will allow the newly started ingester to properly receive a
transfer request or auto join.

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>