Use Ticker for WAL series spreading
Right now series are spread out across the WAL checkpoint duration using
a time.After() call, however that does not account for the amount of
time it may take to checkpoint the series in question. I am consistently
seeing a checkpoint take longer than the goal, and this will help
account for that difference.

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>