Support Multiple Alertmanager URLs
At the moment, the Ruler only support sending alerts to multiple
Alertmanagers via DNS based service discovery. However, sending to
multiple Alertmanager groups is something that Prometheus allows.

I believe the Ruler should support this too and this commit introduces
that.

To keep backward compatibility we're reusing the same flag but allowing
a list of 1+. With this, we can treat each URL as an Alertmanager group
where multiple Alertmanagers per group is only supported if DNS service
discovery is enabled.

Signed-off-by: gotjosh <josue@grafana.com>