fix: `-ruler.alertmanager-url` should be a string
This was never space separated, you needed to register the flag multiple times for it to append to the list of strings.

On top of that, the previous change implied that when using a config file you would need to provide a a list instead of a string thus breaking exisiting configuration.

Signed-off-by: gotjosh <josue@grafana.com>