Simpler instrumentation for non-batched memcached
Previously it would record a `Memcached.Get` operation with a nested
`Memcached.GetMulti`; since they are in one-to-one correspondance this is
pointless.

Skip the outer data collection for the non-batched case, and rename the
batched one to `Memcached.GetBatched` which is more descriptive.

These names are used for metric labels and for tracing span names.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>