Bugfix: Enable TLS server certificate validation by default (#3030)
* Ensure TLS server certificate validation is enabled

Fixes configuration for TLS server validation, TLS skip verify was
enabled for all TLS configurations and prevented validation of server
certificates.

Signed-off-by: Christian Simon <simon@swine.de>

* Improve integration tests

This removes the bash script for CA and certificate generation and
replaces it with some golang library. This allows more control over the
certificates generated. The tls-server-name flags are now used to ensure
we are talking to a server with an appropriate certificate.

Signed-off-by: Christian Simon <simon@swine.de>

* Mount volumes during integration tests with `z`

The 'z' option tells Docker that the volume content will be shared
between containers. Otherwise SELinux prevents file access between the
containers.

Signed-off-by: Christian Simon <simon@swine.de>

* Provide a better error if cert or key is missing

Signed-off-by: Christian Simon <simon@swine.de>

* Add comments where x509 material is sourced from

Signed-off-by: Christian Simon <simon@swine.de>