Sync rules on read path to avoid mismatch
Currently our config update API just writes to Object store. We
periodically sync the data from Object store to in-memory, and the
in-memory state is returned in prometheus-rules API call. This means
after a user updates the config, he won't see it reflect in the UI
for minutes.

This PR will store the last updated timestamp and will sync to disk any
rules that have been updated since. This means we'll be emitting a LIST
call per ruler everytime the user calls the prometheus rules endpoint
but it should be rare enough that this shouldn't materially matter.

Signed-off-by: Goutham Veeramachaneni <gouthamve@gmail.com>