Query the entire ingester range for series.
With https://github.com/prometheus/prometheus/pull/8050 we shouldn't be
paging in chunks for the series API lookups. This would mean much lower
memory utilisation and is similar to doing a range query for that
period with the same query.

With that optimisation, we should allow series API to query all of the
ingester instead of just head block.

Signed-off-by: Goutham Veeramachaneni <gouthamve@gmail.com>