Return 502 if the alertmanager is not available
Often, we find it confusing that the in the unlikely event of:

a) creating a configuration b) then deleting it

or

a) no fallbck config b) making a request to the alertmanager

Our users (and ruler) receive a 404. This is generally, acceptable as
the system might not want to reveal that a configuration for a
particular user exists. However, we deliberately within Cortex do not
make any assumptions wrt authentication/authorization.

I feel like we need to be more indicative and let clients know what
exactly is  is happening. There are many status codes that could
potentially fit the bill so any suggestions are welcome.

According to [http
standard](https://tools.ietf.org/html/rfc7231#section-6.5.8) 502
represents:

    The 502 (Bad Gateway) status code indicates that the server, while
    acting as a gateway or proxy, received an invalid response from an
    inbound server it accessed while attempting to fulfill the request.

Given this is not cacheable by default and represents a state within the
server (a user could create a config to restore the service), it seems
to be like appropriate.

Signed-off-by: gotjosh <josue@grafana.com>