Add Ruler and Alertmanager HTTP API compression
Introduces optional compression for both the Ruler and Alertmanager
config and prometheus-compatible API. It defaults to level 6 compression
which seems like acceptable level - we use the same on in the query
front-end.

In practice, we have noticed that tenants might have a high-number of
rules and alerts which tends to size up to 3MB+ of YAML/JSON being output by
the API endpoints.

While testing this in our environments, we saw up to 91% response
compression.

Signed-off-by: gotjosh <josue@grafana.com>