fix: Stale series with ruler evaluation delay
When ruler evaluation delay is enabled, the upstream rule evaluation
logic adds staleness markers for missing series at the actual timestamp
when the evaluation is run rather than the timestamp when the series was
missing: (see:
https://github.com/prometheus/prometheus/blob/6c56a1faaaad07317ff585bda75b99bdba0517ad/rules/manager.go#L647-L660).

This can cause larger gaps than there actually are, as once the data becomes
available again it can't be written as it would be out-of-order.

This fix shifts those staleness marker back in time by the evaluation
delay set for the individual tenant.

Signed-off-by: Christian Simon <simon@swine.de>
Co-authored-by: Jacob Lisi <jlisi@grafana.com>
Co-authored-by: Marco Pracucci <marco@pracucci.com>
Signed-off-by: Christian Simon <simon@swine.de>