address gosec 301, 302 and 306
gosec reports that directories and files created by cortex allow "others" to access them.
For common deployments of cortex the process should be running as a dedicated user & group,
which removes the need to grant "others" access.

This also shouldn't prevent tools operating on the customer data, as they can run with elevated privileges,
assuming the right user prior to execution.