Fixing ambiguous double negative for ingesters
Fixing ambiguous terminology from double negatives;

"doesn't receive neither this and that". As neither is Not Either, this could suggest that it accepts both concurrently as 'either' binds the two operations together, but it really wants to be suggesting that these operations are mutually exclusive.

Not sure this is the clearest explanation I've ever created on a ticket, but i hope it makes sense.

Reading this section made my eye twitch which may just be my own problem so if you think the current version is valid that's fine by me. I'm not an expert in English grammar :-)

Signed-off-by: Toby Jackson <toby@warmfusion.co.uk>