blocksconvert: optionally reduce timestamp jitter
Run with -builder.timestamp-tolerance=5 to adjust sample timestamps by
up to 5ms to align them to an exact number of seconds apart, thereby
reducing the size of blocks in memory or on disk.

Prometheus started doing this in version 2.22, so this option is most
useful for data gathered from older versions.

See https://github.com/prometheus/prometheus/issues/7846 for background
info.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>