alertmanager/ruler: Include affected path upon config access failure
If the alertmanager config directory contains unexpected or invalid files, mention the problematic file path alongside the underlying error.

Without the path, the problem effectively turns into a landmine, blocking even listing the contents of the bucket, with no indication of where the problem is coming from. For example this was the error on alertmanager startup when a `bucket-index.json.gz` file had gotten into the `alerts/` directory:

> level=error ts=2021-03-03T00:28:49.126516247Z caller=cortex.go:419 msg="module failed" module=alertmanager err="invalid service state: Failed, expected: Running, failure: proto: wrong wireType = 7 for field Templates"

This was originally encountered with alertmanager, but the corresponding ruler storages were updated as well for good measure.

Signed-off-by: Nick Parker <nick@nickbp.com>