Extend alertstore to support read/write of clusterpb.FullState. (#4020)
* Extend alertstore to support read/write of clusterpb.FullState.

Required to support the work for ring-based/sharding replication.

Due to the underlying `clusterpb.FullState` type being an upstream type
defined in alertmanager, it is wrapped in a separate message type
(`alertspb.FullStateDesc`). This can be used to cope with backwards
incompatible changes to the upstream data format, if ever necessary.

Assuming we want to store the state alongside the existing state for the
alertmanager, then it makes sense to extend the alertstore. This avoids
having to plumb through a different storage widget or configuration.

The slight complication is that the existing store uses an object path
of `<prefix>/<user>`, so we have to use a different prefix in order to
use the desired object path of `<prefix>/<user>/<object>`.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>

* Review comments.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>