[compactor] Run blocks cleaner before ring ready.
Yesterday we had a consul issue during a rollout and this led to the
ring check being broken. We weren't alerted until 24hrs later when
queries started failing.

We had an alert on
`cortex_compactor_block_cleanup_last_successful_run_timestamp_seconds`
but it wasn't even being set.

This makes sure it is being set before we hit the ring.

Signed-off-by: Goutham Veeramachaneni <gouthamve@gmail.com>