Fix for alertmanagers not configuring users before becoming ACTIVE. (#4110)
* Fix for alertmanagers not configuring users before becoming ACTIVE.

(Only applies to sharding operation)

Currently when the alertmanager joins the sharding ring, it tries to
load user configurations, and checks the ring to see what users it
should be servicing. Once finished, the alertmanager changes it's state
to ACTIVE.

However, when performing this initial sync of user configurations, the
alertmanager is in the JOINING state, but checking if a user is owned
requires an instance to be in the ACTIVE state. Therefore, the initial
sync will _never_ configure any users, but the instance will be
considered ACTIVE.

This change fixes the initial sync by considering a user to be owned by
the instance if it is in either the ACTIVE or JOINING state.

Note this fix should address the sporadic integration test failures:
The distributor forwards requests to an instance which it believes is
configured, but the request then fails, because it is not ready yet.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>

* Fix race condition in unit test.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>

* Review comments.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>