First cut at metrics for alertmanager sharding operation.
This commit adds a number of metrics to help track the operation of the
alertmanager sharding, specifically around the handling of the state.

- `cortex_alertmanager_fetch_replica_state_total`
- `cortex_alertmanager_fetch_replica_state_failed_total`
- `cortex_alertmanager_state_initial_sync_total`
- `cortex_alertmanager_state_initial_sync_completed_total`
- `cortex_alertmanager_state_initial_sync_duration_seconds`
- `cortex_alertmanager_state_persist_total`
- `cortex_alertmanager_state_persist_failed_total`

Note this complements the already existing metrics:

- `cortex_alertmanager_partial_state_merges_total`
- `cortex_alertmanager_partial_state_merges_failed_total`
- `cortex_alertmanager_state_replication_total`
- `cortex_alertmanager_state_replication_failed_total`

Overly detailed timing metrics have not been included, instead opting
for just a single (non per-user) histogram for the duration of the
initial state operation. Timings for storage read/write are not
included as they are already provided from the bucket client.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>