Make handler function for `/config` endpoint overrideable
In order to be able to extend or replace the default handler function for the
`/config` endpoint, the registered handler evaluates whether a handler override
is set. If so, the custom handler function is invoked, otherwise the default
handler is used.

An exampe for this use-case is, that a downstream project extends the Cortex
configuration with additional configuration options, which should also be
exposed in the `/config` endpoint.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>