Enable active series metrics in the ingester by default
This change calculates and exports the `cortex_ingester_active_series`
by default. Up to this point, the metric was disabled by default since
calculating it consumes some amount of memory.

The original PR (#3153) estimated for 1M active series at least 40MB
and up to another 200MB depending on our luck reusing labels from the
ref cache.

We (Grafana) have been running with this setting enabled on our ingesters
for some time and the resource usage doesn't appear to be significant.
This feature appears to add between 1.2 - 1.6% in memory usage when
enabled: ~140MB out of a total of ~10GB of memory used per ingester.

The ingesters I measured this on

* Have multiple tenants running production workloads
* Have about 1.3M active series each
* Have about a 10GB working set (as measured by `kubectl top` and exported
  k8s metrics)

Based on this and the utility of the metric itself, I'd like to enable it
by default.

Screenshots of the pprof heap output attached.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>