Reuse the byte buffer from GRPC response in the frontend.
This PR allow to reuse the GRPC byte buffer casted as a `io.Reader` in the http rountripper of the frontend.
This way we don't have to copy the content from the reader into another buffer when reading the http response.

I've been testing this with in Loki and this shows good memory saving (around 1GB).

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>