Upgrade Prometheus to LabelNames with matchers
Since https://github.com/prometheus/prometheus/pull/9083 prometheus now
provides matchers to the LabelNames method on the LabelQuerier
interface, so in order to upgrade Prometheus we need to support that.

This partially solves
https://github.com/cortexproject/cortex/issues/3658 as now the block
store queryable uses the LabelNames method with matchers.

However, we're still using the ingesters' MetricsForLabelMatchers method
to perform the LabelNames w/matchers call on the distributor. That
change will be tackled separately as it breaks ingester's contracts and
needs to be released with a feature flag to perform a backwards
compatible release.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>