Reduce flakiness on chunk flush tests (#4386)
The thinking was: we are flushing every 20ms, so wait 40ms to
be sure that one cycle has happened. However if you log the times at
which things happen, it is fairly common to see 100ms delay on the flush
ticker. So wait a lot longer to be sure.

Also close down the ingester that we created for the test, and improve
the error message on failure.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>
Signed-off-by: Alvin Lin <alvinlin@amazon.com>