Fix race when stopping chunks memcached client
We need to check again if we have been asked to quit before writing to
the results chan.

Also, the reader of that chan should not close it, because this can
trigger a panic in a writer. The chan will be cleaned up by garbage-
collection once all the readers and writers have exited on quit.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>