ingester: add max active series per user metric.
Max active series per user depends on default and per-tenant limits, as
well as distributor sharding strategy and number of healthy ingesters.
To simplify alerting when a tenant is approaching this limit, add a
`cortex_ingester_max_active_series` metric that is updated when the
existing `cortex_ingester_active_series` changes.

Signed-off-by: Josh Carp <jm.carp@gmail.com>