ring: Restart initRing with unexpected token count
This commit updates the ingester initRing logic so that the number of
tokens per ingester stated in configuration is obeyed. Currently, when
an ingester starts up, it blindly accepts the tokens pulled from the
ring for its ID. Thus the counts between configuration and what is found
in the ring can be different.

Fixes #1876.

Signed-off-by: Paul Logston <logston@cockroachlabs.com>