Update(ADOPTERS.md): add "us" (#5874)
For context, we've been using Cortex since 2020/2021 in production. It's the backbone for all the metrics we provide to customers ourselves and also for custom integrations (where customers run their own exporters, etc..).

Just saw the cnf issue about the project, and wanted to add us. We are not planning to switch to Mimir any time soon. :)

Signed-off-by: Till! <till@users.noreply.github.com>