Select correct tenant during query federation
We are creating a job which has the tenant name for each tenant to
query, but we use the incorrect index in the `ids` slice that contains
the list of tenants. This causes the querier to select results for the
first tenant in the list of tenants to match with instead of the actual
tenant the querier was intended for.

I updated the existing tests to check the labels of the results as well
instead of only relying on the number of series found.

Fixes https://github.com/cortexproject/cortex/issues/5941

Signed-off-by: Charlie Le <charlie_le@apple.com>