Wait until container is removed after stopping
The E2E tests use `docker stop` to stop the container, but the container
can still be in the process of being removed after the stop command is
executed. This can cause tests that stop and start a container in rapid
succession with the same container name to fail with conflicting
container names.

Example error message in an integration test:

```
00:08:14 Stopping query-frontend
00:08:14 Stopping store-gateway
=== RUN
TestNewDistributorsCanPushToOldIngestersWithReplication/Backward_compatibility_upgradir
00:08:14 Starting store-gateway
00:08:14 store-gateway: docker: Error response from daemon: Conflict. The container name "/e2e-cortex-test-store-gateway" is already in use by container "b750a7b7e2e3eebc617ac832de2c8a643f884bbc79b0c8961ceebace8cad5c85". You have to remove (or rename) that container to be able to reuse that name.
```

Signed-off-by: Charlie Le <charlie_le@apple.com>